+++ src/NetDevPack.Security.Jwt.Core/Model/RevokedKeyInfo.cs
@@ -1,0 +1,7 @@
+﻿namespace NetDevPack.Security.Jwt.Core.Model;
+
+record RevokedKeyInfo(string Id, string? RevokedReason=default)
+{
+    public string Id { get; } = Id;
+    public string? RevokedReason { get; } = RevokedReason;
+}